Added package statement and corrected root samples folder name
