--- sdk/openai/azure-ai-openai/src/sample/java/com.azure.openai.samples/ResponsesImageUrlExample.java
@@ -1,56 +1,0 @@
-import com.azure.identity.AuthenticationUtil;
-import com.azure.identity.DefaultAzureCredentialBuilder;
-import com.openai.azure.AzureOpenAIServiceVersion;
-import com.openai.client.OpenAIClient;
-import com.openai.client.okhttp.OpenAIOkHttpClient;
-import com.openai.credential.BearerTokenCredential;
-import com.openai.models.ChatModel;
-import com.openai.models.responses.ResponseCreateParams;
-import com.openai.models.responses.ResponseInputImage;
-import com.openai.models.responses.ResponseInputItem;
-
-import java.io.IOException;
-import java.util.Base64;
-import java.util.List;
-
-public final class ResponsesImageUrlExample {
-    private ResponsesImageUrlExample() {}
-
-    public static void main(String[] args) throws IOException {
-        // Configures using one of:
-        // - The `OPENAI_API_KEY` environment variable
-        // - The `AZURE_OPENAI_ENDPOINT` and `AZURE_OPENAI_KEY` environment variables
-        OpenAIClient client = OpenAIOkHttpClient.builder()
-            .azureServiceVersion(AzureOpenAIServiceVersion.latestPreviewVersion())
-            // Gets the API key from the `AZURE_OPENAI_KEY` environment variable
-            .baseUrl(System.getenv("AZURE_OPENAI_ENDPOINT"))
-            // Set the Azure Entra ID
-            .credential(BearerTokenCredential.create(AuthenticationUtil.getBearerTokenSupplier(
-                new DefaultAzureCredentialBuilder().build(), "https://cognitiveservices.azure.com/.default")))
-            .build();
-
-        ClassLoader classloader = Thread.currentThread().getContextClassLoader();
-        byte[] logoBytes = classloader.getResource("logo.png").openStream().readAllBytes();
-        String logoBase64Url = "data:image/jpeg;base64," + Base64.getEncoder().encodeToString(logoBytes);
-
-        ResponseInputImage logoInputImage = ResponseInputImage.builder()
-                .detail(ResponseInputImage.Detail.AUTO)
-                .imageUrl(logoBase64Url)
-                .build();
-        ResponseInputItem messageInputItem = ResponseInputItem.ofMessage(ResponseInputItem.Message.builder()
-                .role(ResponseInputItem.Message.Role.USER)
-                .addInputTextContent("Describe this image.")
-                .addContent(logoInputImage)
-                .build());
-        ResponseCreateParams createParams = ResponseCreateParams.builder()
-                .inputOfResponse(List.of(messageInputItem))
-                .model(ChatModel.GPT_4O_MINI)
-                .build();
-
-        client.responses().create(createParams).output().stream()
-                .flatMap(item -> item.message().stream())
-                .flatMap(message -> message.content().stream())
-                .flatMap(content -> content.outputText().stream())
-                .forEach(outputText -> System.out.println(outputText.text()));
-    }
-}
