+++ sdk/openai/azure-ai-openai/src/samples/java/com.azure.openai.samples/CompletionsExample.java
@@ -1,21 +1,27 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.openai.samples;
+
 import com.azure.identity.AuthenticationUtil;
 import com.azure.identity.DefaultAzureCredentialBuilder;
-import com.openai.client.OpenAIClientAsync;
-import com.openai.client.okhttp.OpenAIOkHttpClientAsync;
-import com.openai.core.http.AsyncStreamResponse;
+import com.openai.client.OpenAIClient;
+import com.openai.client.okhttp.OpenAIOkHttpClient;
 import com.openai.credential.BearerTokenCredential;
 import com.openai.models.ChatModel;
-import com.openai.models.chat.completions.ChatCompletionChunk;
 import com.openai.models.chat.completions.ChatCompletionCreateParams;
 
-public final class CompletionsStreamingCancellationAsyncExample {
-    private CompletionsStreamingCancellationAsyncExample() {}
+public final class CompletionsExample {
+    private CompletionsExample() {}
 
     public static void main(String[] args) {
         // Configures using one of:
         // - The `OPENAI_API_KEY` environment variable
         // - The `AZURE_OPENAI_ENDPOINT` and `AZURE_OPENAI_KEY` environment variables
-        OpenAIOkHttpClientAsync.Builder clientBuilder = OpenAIOkHttpClientAsync.builder();
+        // Configures using one of:
+        // - The `OPENAI_API_KEY` environment variable
+        // - The `AZURE_OPENAI_ENDPOINT` and `AZURE_OPENAI_KEY` environment variables
+        OpenAIOkHttpClient.Builder clientBuilder = OpenAIOkHttpClient.builder();
 
         /* Azure-specific code starts here */
         // You can either set 'endpoint' or 'apiKey' directly in the builder.
@@ -27,9 +33,7 @@ public final class CompletionsStreamingCancellationAsyncExample {
         /* Azure-specific code ends here */
 
         // All code from this line down is general-purpose OpenAI code
-        OpenAIClientAsync client = clientBuilder.build();
-
-
+        OpenAIClient client = clientBuilder.build();
         ChatCompletionCreateParams createParams = ChatCompletionCreateParams.builder()
                 .model(ChatModel.GPT_4O)
                 .maxCompletionTokens(2048)
@@ -37,19 +41,8 @@ public final class CompletionsStreamingCancellationAsyncExample {
                 .addUserMessage("Tell me a story about building the best SDK!")
                 .build();
 
-        AsyncStreamResponse<ChatCompletionChunk> streamResponse =
-                client.chat().completions().createStreaming(createParams);
-        streamResponse
-                .subscribe(completion -> completion.choices().stream()
-                        .flatMap(choice -> choice.delta().content().stream())
-                        .forEach(text -> {
-                            System.out.print(text);
-                            if (text.contains("SDK")) {
-                                // Close the stream early.
-                                streamResponse.close();
-                            }
-                        }))
-                .onCompleteFuture()
-                .join();
+        client.chat().completions().create(createParams).choices().stream()
+                .flatMap(choice -> choice.message().content().stream())
+                .forEach(System.out::println);
     }
 }
