+++ sdk/openai/azure-ai-openai/src/samples/java/com.azure.openai.samples/CompletionsStreamingExample.java
@@ -1,13 +1,20 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.openai.samples;
+
 import com.azure.identity.AuthenticationUtil;
 import com.azure.identity.DefaultAzureCredentialBuilder;
 import com.openai.client.OpenAIClient;
 import com.openai.client.okhttp.OpenAIOkHttpClient;
+import com.openai.core.http.StreamResponse;
 import com.openai.credential.BearerTokenCredential;
-import com.openai.models.embeddings.EmbeddingCreateParams;
-import com.openai.models.embeddings.EmbeddingModel;
+import com.openai.models.ChatModel;
+import com.openai.models.chat.completions.ChatCompletionChunk;
+import com.openai.models.chat.completions.ChatCompletionCreateParams;
 
-public final class EmbeddingsExample {
-    private EmbeddingsExample() {}
+public final class CompletionsStreamingExample {
+    private CompletionsStreamingExample() {}
 
     public static void main(String[] args) {
         // Configures using one of:
@@ -27,11 +34,19 @@ public final class EmbeddingsExample {
         // All code from this line down is general-purpose OpenAI code
         OpenAIClient client = clientBuilder.build();
 
-        EmbeddingCreateParams createParams = EmbeddingCreateParams.builder()
-                .input("The quick brown fox jumped over the lazy dog")
-                .model(EmbeddingModel.TEXT_EMBEDDING_ADA_002)
+        ChatCompletionCreateParams createParams = ChatCompletionCreateParams.builder()
+                .model(ChatModel.GPT_4O)
+                .maxCompletionTokens(2048)
+                .addDeveloperMessage("Make sure you mention Stainless!")
+                .addUserMessage("Tell me a story about building the best SDK!")
                 .build();
 
-        System.out.println(client.embeddings().create(createParams));
+        try (StreamResponse<ChatCompletionChunk> streamResponse =
+                client.chat().completions().createStreaming(createParams)) {
+            streamResponse.stream()
+                    .flatMap(completion -> completion.choices().stream())
+                    .flatMap(choice -> choice.delta().content().stream())
+                    .forEach(System.out::print);
+        }
     }
 }
