+++ sdk/openai/azure-ai-openai/src/samples/java/com.azure.openai.samples/EmbeddingsAsyncExample.java
@@ -1,18 +1,20 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.openai.samples;
+
 import com.azure.identity.AuthenticationUtil;
 import com.azure.identity.DefaultAzureCredentialBuilder;
 import com.openai.client.OpenAIClientAsync;
 import com.openai.client.okhttp.OpenAIOkHttpClientAsync;
 import com.openai.credential.BearerTokenCredential;
-import com.openai.models.audio.AudioModel;
-import com.openai.models.audio.transcriptions.TranscriptionCreateParams;
+import com.openai.models.embeddings.EmbeddingCreateParams;
+import com.openai.models.embeddings.EmbeddingModel;
 
-import java.nio.file.Path;
-import java.nio.file.Paths;
+public final class EmbeddingsAsyncExample {
+    private EmbeddingsAsyncExample() {}
 
-public final class AudioTranscriptionsAsyncExample {
-    private AudioTranscriptionsAsyncExample() {}
-
-    public static void main(String[] args) throws Exception {
+    public static void main(String[] args) {
         // Configures using one of:
         // - The `OPENAI_API_KEY` environment variable
         // - The `AZURE_OPENAI_ENDPOINT` and `AZURE_OPENAI_KEY` environment variables
@@ -30,20 +32,11 @@ public final class AudioTranscriptionsAsyncExample {
         // All code from this line down is general-purpose OpenAI code
         OpenAIClientAsync client = clientBuilder.build();
 
-
-        ClassLoader classloader = Thread.currentThread().getContextClassLoader();
-        Path path = Paths.get(classloader.getResource("sports.wav").toURI());
-
-        TranscriptionCreateParams createParams = TranscriptionCreateParams.builder()
-                .file(path)
-                .model(AudioModel.of("whisper"))
+        EmbeddingCreateParams createParams = EmbeddingCreateParams.builder()
+                .input("The quick brown fox jumped over the lazy dog")
+                .model(EmbeddingModel.TEXT_EMBEDDING_ADA_002)
                 .build();
 
-        client.audio()
-                .transcriptions()
-                .create(createParams)
-                .thenAccept(response ->
-                        System.out.println(response.asTranscription().text()))
-                .join();
+        client.embeddings().create(createParams).thenAccept(System.out::println).join();
     }
 }
