+++ sdk/openai/azure-ai-openai/src/samples/java/com.azure.openai.samples/Main.java
@@ -1,3 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.openai.samples;
 
 public final class Main {
     private Main() {}
