+++ sdk/openai/azure-ai-openai/src/samples/java/com.azure.openai.samples/ResponsesExample.java
@@ -1,22 +1,25 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.openai.samples;
+
 import com.azure.identity.AuthenticationUtil;
 import com.azure.identity.DefaultAzureCredentialBuilder;
 import com.openai.azure.AzureOpenAIServiceVersion;
 import com.openai.client.OpenAIClient;
-import com.openai.client.OpenAIClientAsync;
 import com.openai.client.okhttp.OpenAIOkHttpClient;
-import com.openai.client.okhttp.OpenAIOkHttpClientAsync;
 import com.openai.credential.BearerTokenCredential;
 import com.openai.models.ChatModel;
 import com.openai.models.responses.ResponseCreateParams;
 
-public final class ResponsesStreamingAsyncExample {
-    private ResponsesStreamingAsyncExample() {}
+public final class ResponsesExample {
+    private ResponsesExample() {}
 
     public static void main(String[] args) {
         // Configures using one of:
         // - The `OPENAI_API_KEY` environment variable
         // - The `AZURE_OPENAI_ENDPOINT` and `AZURE_OPENAI_KEY` environment variables
-        OpenAIClientAsync client = OpenAIOkHttpClientAsync.builder()
+        OpenAIClient client = OpenAIOkHttpClient.builder()
             .azureServiceVersion(AzureOpenAIServiceVersion.latestPreviewVersion())
             // Gets the API key from the `AZURE_OPENAI_KEY` environment variable
             .baseUrl(System.getenv("AZURE_OPENAI_ENDPOINT"))
@@ -30,10 +33,10 @@ public final class ResponsesStreamingAsyncExample {
                 .model(ChatModel.GPT_4O_MINI)
                 .build();
 
-        client.responses()
-                .createStreaming(createParams)
-                .subscribe(event -> event.outputTextDelta().ifPresent(textEvent -> System.out.print(textEvent.delta())))
-                .onCompleteFuture()
-                .join();
+        client.responses().create(createParams).output().stream()
+                .flatMap(item -> item.message().stream())
+                .flatMap(message -> message.content().stream())
+                .flatMap(content -> content.outputText().stream())
+                .forEach(outputText -> System.out.println(outputText.text()));
     }
 }
