+++ sdk/openai/azure-ai-openai/src/samples/java/com.azure.openai.samples/ResponsesImageUrlExample.java
@@ -1,3 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.openai.samples;
+
 import com.azure.identity.AuthenticationUtil;
 import com.azure.identity.DefaultAzureCredentialBuilder;
 import com.openai.azure.AzureOpenAIServiceVersion;
@@ -6,11 +11,17 @@ import com.openai.client.okhttp.OpenAIOkHttpClient;
 import com.openai.credential.BearerTokenCredential;
 import com.openai.models.ChatModel;
 import com.openai.models.responses.ResponseCreateParams;
+import com.openai.models.responses.ResponseInputImage;
+import com.openai.models.responses.ResponseInputItem;
 
-public final class ResponsesExample {
-    private ResponsesExample() {}
+import java.io.IOException;
+import java.util.Base64;
+import java.util.List;
 
-    public static void main(String[] args) {
+public final class ResponsesImageUrlExample {
+    private ResponsesImageUrlExample() {}
+
+    public static void main(String[] args) throws IOException {
         // Configures using one of:
         // - The `OPENAI_API_KEY` environment variable
         // - The `AZURE_OPENAI_ENDPOINT` and `AZURE_OPENAI_KEY` environment variables
@@ -23,8 +34,21 @@ public final class ResponsesExample {
                 new DefaultAzureCredentialBuilder().build(), "https://cognitiveservices.azure.com/.default")))
             .build();
 
+        ClassLoader classloader = Thread.currentThread().getContextClassLoader();
+        byte[] logoBytes = classloader.getResource("logo.png").openStream().readAllBytes();
+        String logoBase64Url = "data:image/jpeg;base64," + Base64.getEncoder().encodeToString(logoBytes);
+
+        ResponseInputImage logoInputImage = ResponseInputImage.builder()
+                .detail(ResponseInputImage.Detail.AUTO)
+                .imageUrl(logoBase64Url)
+                .build();
+        ResponseInputItem messageInputItem = ResponseInputItem.ofMessage(ResponseInputItem.Message.builder()
+                .role(ResponseInputItem.Message.Role.USER)
+                .addInputTextContent("Describe this image.")
+                .addContent(logoInputImage)
+                .build());
         ResponseCreateParams createParams = ResponseCreateParams.builder()
-                .input("Tell me a story about building the best SDK!")
+                .inputOfResponse(List.of(messageInputItem))
                 .model(ChatModel.GPT_4O_MINI)
                 .build();
