+++ sdk/openai/azure-ai-openai/src/samples/java/com.azure.openai.samples/ResponsesStreamingAsyncExample.java
@@ -1,22 +1,27 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.openai.samples;
+
 import com.azure.identity.AuthenticationUtil;
 import com.azure.identity.DefaultAzureCredentialBuilder;
 import com.openai.azure.AzureOpenAIServiceVersion;
 import com.openai.client.OpenAIClient;
+import com.openai.client.OpenAIClientAsync;
 import com.openai.client.okhttp.OpenAIOkHttpClient;
-import com.openai.core.http.StreamResponse;
+import com.openai.client.okhttp.OpenAIOkHttpClientAsync;
 import com.openai.credential.BearerTokenCredential;
 import com.openai.models.ChatModel;
 import com.openai.models.responses.ResponseCreateParams;
-import com.openai.models.responses.ResponseStreamEvent;
 
-public final class ResponsesStreamingExample {
-    private ResponsesStreamingExample() {}
+public final class ResponsesStreamingAsyncExample {
+    private ResponsesStreamingAsyncExample() {}
 
     public static void main(String[] args) {
         // Configures using one of:
         // - The `OPENAI_API_KEY` environment variable
         // - The `AZURE_OPENAI_ENDPOINT` and `AZURE_OPENAI_KEY` environment variables
-        OpenAIClient client = OpenAIOkHttpClient.builder()
+        OpenAIClientAsync client = OpenAIOkHttpClientAsync.builder()
             .azureServiceVersion(AzureOpenAIServiceVersion.latestPreviewVersion())
             // Gets the API key from the `AZURE_OPENAI_KEY` environment variable
             .baseUrl(System.getenv("AZURE_OPENAI_ENDPOINT"))
@@ -27,14 +32,13 @@ public final class ResponsesStreamingExample {
 
         ResponseCreateParams createParams = ResponseCreateParams.builder()
                 .input("Tell me a story about building the best SDK!")
-                .model(ChatModel.GPT_4O)
+                .model(ChatModel.GPT_4O_MINI)
                 .build();
 
-        try (StreamResponse<ResponseStreamEvent> streamResponse =
-                client.responses().createStreaming(createParams)) {
-            streamResponse.stream()
-                    .flatMap(event -> event.outputTextDelta().stream())
-                    .forEach(textEvent -> System.out.print(textEvent.delta()));
-        }
+        client.responses()
+                .createStreaming(createParams)
+                .subscribe(event -> event.outputTextDelta().ifPresent(textEvent -> System.out.print(textEvent.delta())))
+                .onCompleteFuture()
+                .join();
     }
 }
